unify timer stuff

+++ kernel/oswrapper/timer.h
@@ -11,8 +11,8 @@ EXTERN_VAR int timerv;
 void startTimer(void);
 void writeTime(const char* s);
 
-int initTimer();
-int  getTimer();
+int  initTimer();
+long getTimer();
 
 EXTERN_VAR int rtimerv;
 void startRTimer(void);
